+++ lib/philiprehberger/circuit/state.rb
@@ -5,5 +5,7 @@ module Philiprehberger
     CLOSED    = :closed
     OPEN      = :open
     HALF_OPEN = :half_open
+
+    STATE_CALLBACK_MAP = { CLOSED => :close, OPEN => :open, HALF_OPEN => :half_open }.freeze
   end
 end